Add test cases for adding the game and listing the game methods
