+++ spec/game_spec.rb
@@ -1,4 +1,5 @@
 require 'date'
+require 'stringio'
 require_relative 'spec_helper'
 
 describe Game do
@@ -30,11 +31,19 @@ describe Game do
     end
   end
 
-  describe 'add_game' do
-    
+  describe GameModule do
+    it 'add new game to the game list' do
+      dc = App.new
+      puts @games
+      expect(dc.add_game("true", "2021-07-07", "2011-04-04", "Mike", "Atherton")).to eq('Game has been created successfully')
+    end
   end
 
-  describe 'list_games' do
-
+  describe GameModule do
+    it 'Number of item in list the games equal to 1' do
+      dc = App.new
+      dc.add_game("true", "2021-07-07", "2011-04-04", "Mike", "Atherton")
+      expect(dc.list_games.size).to eq(1)
+    end
   end
 end
